Continue simplifying testhelper
